_create_response_messageをファイルに分割しました
